Сервер выведен в отдельный класс
