+++ src/Handlers/EpicsHandler.java
@@ -1,6 +1,5 @@
 package Handlers;
 
-import Handlers.TypeTokens.EpicsToken;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
@@ -9,11 +8,7 @@ import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import task.Epic;
 import task.Subtask;
-import task.Task;
-import taskmanager.FileBackedTaskManager;
 import taskmanager.TaskManager;
-
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
@@ -21,10 +16,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class EpicsHandler implements HttpHandler {
+public class EpicsHandler extends BaseHttpHandler implements HttpHandler {
 
     TaskManager manager;
-    BaseHttpHandler baseHttpHandler = new BaseHttpHandler();
     public EpicsHandler(TaskManager manager) {
         this.manager = manager;
     }
@@ -33,9 +27,9 @@ public class EpicsHandler implements HttpHandler {
     @Override
     public void handle(HttpExchange exchange) {
         try {
-            Gson gson = baseHttpHandler.createGson();
+            Gson gson = createGson();
             String method = exchange.getRequestMethod();
-            int pathLength = baseHttpHandler.getPathLength(exchange);
+            int pathLength = getPathLength(exchange);
 
 
             switch (method) {
@@ -44,34 +38,34 @@ public class EpicsHandler implements HttpHandler {
                         case 2 -> {
                             List<Epic> tasks = manager.getEpicTasks();
                             if (tasks == null) {
-                                baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                sendNotFound(exchange, "Задача не найдена");
                             }
                             String tasksJson = gson.toJson(tasks);
-                            baseHttpHandler.sendText(exchange, tasksJson);
+                            sendText(exchange, tasksJson);
                         }
                         case 3 -> {
-                            Optional<Integer> id = baseHttpHandler.getTaskId(exchange);
+                            Optional<Integer> id = getTaskId(exchange);
                             if (id.isEmpty()) {
-                                baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                sendNotFound(exchange, "Задача не найдена");
                             } else {
                                 Epic newTask = manager.getEpicTask(id.get());
                                 if (newTask == null) {
-                                    baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                    sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
                             Epic task = manager.getEpicTask(id.get());
                             String taskJson = gson.toJson(task);
-                            baseHttpHandler.sendText(exchange, taskJson);
+                            sendText(exchange, taskJson);
                         }
                         case 4 -> {
-                            Optional<Integer> epicId = baseHttpHandler.getTaskId(exchange);
+                            Optional<Integer> epicId = getTaskId(exchange);
                             if (epicId.isEmpty()) {
-                                baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                sendNotFound(exchange, "Задача не найдена");
                             } else {
                                 if (manager.getEpicTask(epicId.get()) != null) {
                                     ArrayList<Subtask> subtasks = manager.getEpicSubtasks(epicId.get());
                                     String subtasksJson = gson.toJson(subtasks);
-                                    baseHttpHandler.sendText(exchange, subtasksJson);
+                                    sendText(exchange, subtasksJson);
                                 }
                             }
                         }
@@ -82,7 +76,7 @@ public class EpicsHandler implements HttpHandler {
                     String stringInput = new String(input.readAllBytes(), StandardCharsets.UTF_8);
                     JsonElement element = JsonParser.parseString(stringInput);
                     if (!element.isJsonObject()) {
-                        baseHttpHandler.sendNotAcceptable(exchange, "Это не Json объект");
+                        sendNotAcceptable(exchange, "Это не Json объект");
                     }
                     JsonObject jsonObject = element.getAsJsonObject();
                     Epic taskFromJson = gson.fromJson(jsonObject, Epic.class);
@@ -93,17 +87,17 @@ public class EpicsHandler implements HttpHandler {
                             exchange.close();
                         }
                         case 3 -> {
-                            Optional<Integer> id = baseHttpHandler.getTaskId(exchange);
+                            Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
                                 Epic task2 = manager.updateEpicTask(taskFromJson);
                                 if (task2 != null) {
                                         exchange.sendResponseHeaders(201, 0);
                                         exchange.close();
                                 } else {
-                                    baseHttpHandler.sendNotAcceptable(exchange, "Возникло пересечение");
+                                    sendNotAcceptable(exchange, "Возникло пересечение");
                                 }
                             } else {
-                                baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                sendNotFound(exchange, "Задача не найдена");
                             }
                         }
                     }
@@ -117,14 +111,14 @@ public class EpicsHandler implements HttpHandler {
 
                         }
                         case 3 -> {
-                            Optional<Integer> id = baseHttpHandler.getTaskId(exchange);
+                            Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
                                 if (manager.getEpicTask(id.get()) != null) {
                                     manager.deleteEpic(id.get());
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
                                 } else {
-                                    baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                    sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
                         }
