+++ src/Handlers/TasksHandler.java
@@ -1,28 +1,19 @@
 package Handlers;
 
-import Handlers.TypeTokens.TasksToken;
-import category.TaskStatus;
 import com.google.gson.*;
-import com.google.gson.internal.bind.util.ISO8601Utils;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import task.Task;
-import taskmanager.FileBackedTaskManager;
 import taskmanager.TaskManager;
-
 import java.io.*;
-
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
-import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class TasksHandler implements HttpHandler {
+public class TasksHandler extends BaseHttpHandler implements HttpHandler {
 
     TaskManager manager;
-    BaseHttpHandler baseHttpHandler = new BaseHttpHandler();
 
     public TasksHandler(TaskManager manager) {
         this.manager = manager;
@@ -32,9 +23,9 @@ public class TasksHandler implements HttpHandler {
     @Override
     public void handle(HttpExchange exchange) {
         try {
-            Gson gson = baseHttpHandler.createGson();
+            Gson gson = createGson();
             String method = exchange.getRequestMethod();
-            int pathLength = baseHttpHandler.getPathLength(exchange);
+            int pathLength = getPathLength(exchange);
 
 
             switch (method) {
@@ -43,24 +34,24 @@ public class TasksHandler implements HttpHandler {
                         case 2 -> {
                             List<Task> tasks = manager.getSimpleTasks();
                             if (tasks == null) {
-                                baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                sendNotFound(exchange, "Задача не найдена");
                             }
                             String tasksJson = gson.toJson(tasks);
-                            baseHttpHandler.sendText(exchange, tasksJson);
+                            sendText(exchange, tasksJson);
                         }
                         case 3 -> {
-                            Optional<Integer> id = baseHttpHandler.getTaskId(exchange);
+                            Optional<Integer> id = getTaskId(exchange);
                             if (id.isEmpty()) {
-                                baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                sendNotFound(exchange, "Задача не найдена");
                             } else {
                                 Task newTask = manager.getSimpleTask(id.get());
                                 if (newTask == null) {
-                                    baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                    sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
                             Task task = manager.getSimpleTask(id.get());
                             String taskJson = gson.toJson(task);
-                            baseHttpHandler.sendText(exchange, taskJson);
+                            sendText(exchange, taskJson);
                         }
                     }
                 }
@@ -69,7 +60,7 @@ public class TasksHandler implements HttpHandler {
                     String stringInput = new String(input.readAllBytes(), StandardCharsets.UTF_8);
                     JsonElement element = JsonParser.parseString(stringInput);
                     if (!element.isJsonObject()) {
-                        baseHttpHandler.sendNotAcceptable(exchange, "Это не Json объект");
+                        sendNotAcceptable(exchange, "Это не Json объект");
                     }
                     JsonObject jsonObject = element.getAsJsonObject();
                     Task taskFromJson = gson.fromJson(jsonObject, Task.class);
@@ -77,24 +68,24 @@ public class TasksHandler implements HttpHandler {
                         case 2 -> {
                             Task newTask = manager.addSimpleTask(taskFromJson);
                             if (newTask == null) {
-                                baseHttpHandler.sendNotAcceptable(exchange, "Задача пересекается по времени");
+                                sendNotAcceptable(exchange, "Задача пересекается по времени");
                             } else {
                                 exchange.sendResponseHeaders(201, 0);
                                 exchange.close();
                             }
                         }
                         case 3 -> {
-                            Optional<Integer> id = baseHttpHandler.getTaskId(exchange);
+                            Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
                                 Task task2 = manager.updateSimpleTask(taskFromJson);
                                 if (task2 != null) {
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
                                 } else {
-                                    baseHttpHandler.sendNotAcceptable(exchange, "Возникло пересечение");
+                                    sendNotAcceptable(exchange, "Возникло пересечение");
                                 }
                             } else {
-                                baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                sendNotFound(exchange, "Задача не найдена");
                             }
                         }
                     }
@@ -107,14 +98,14 @@ public class TasksHandler implements HttpHandler {
                             exchange.close();
                         }
                         case 3 -> {
-                            Optional<Integer> id = baseHttpHandler.getTaskId(exchange);
+                            Optional<Integer> id = getTaskId(exchange);
                             if (id.isPresent()) {
                                 if (manager.getSimpleTask(id.get()) != null) {
                                     manager.deleteSimpleTask(id.get());
                                     exchange.sendResponseHeaders(201, 0);
                                     exchange.close();
                                 } else {
-                                    baseHttpHandler.sendNotFound(exchange, "Задача не найдена");
+                                    sendNotFound(exchange, "Задача не найдена");
                                 }
                             }
                         }
