+++ test/HandlersTests/TasksHandlerTest.java
@@ -1,0 +1,66 @@
+package HandlersTests;
+
+import Handlers.BaseHttpHandler;
+import com.google.gson.Gson;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.AfterEach;
+import taskmanager.InMemoryTaskManager;
+import taskmanager.TaskManager;
+
+import java.io.IOException;
+import java.net.URI;
+import java.net.http.HttpClient;
+import java.net.http.HttpRequest;
+import java.net.http.HttpResponse;
+
+public class TasksHandlerTest extends BaseHttpHandler {
+
+    // создаём экземпляр InMemoryTaskManager
+    TaskManager manager = new InMemoryTaskManager();
+    // передаём его в качестве аргумента в конструктор HttpTaskServer
+    BaseHttpHandler baseHandler = new BaseHttpHandler();
+    Gson gson = baseHandler
+
+    public TasksHandlerTest() throws IOException {
+    }
+
+    @BeforeEach
+    public void setUp() {
+        manager.deleteTasks();
+        manager.deleteSubtasks();
+        manager.deleteEpics();
+        taskServer.start();
+    }
+
+    @AfterEach
+    public void shutDown() {
+        taskServer.stop();
+    }
+
+    @Test
+    public void testAddTask() throws IOException, InterruptedException {
+        // создаём задачу
+        Task task = new Task("Test 2", "Testing task 2",
+                TaskStatus.NEW, Duration.ofMinutes(5), LocalDateTime.now());
+        // конвертируем её в JSON
+        String taskJson = gson.toJson(task);
+
+        // создаём HTTP-клиент и запрос
+        HttpClient client = HttpClient.newHttpClient();
+        URI url = URI.create("http://localhost:8080/tasks");
+        HttpRequest request = HttpRequest.newBuilder().uri(url).POST(HttpRequest.BodyPublishers.ofString(taskJson)).build();
+
+        // вызываем рест, отвечающий за создание задач
+        HttpResponse<String> response = client.send(request, HttpResponse.BodyHandlers.ofString());
+        // проверяем код ответа
+        assertEquals(200, response.statusCode());
+
+        // проверяем, что создалась одна задача с корректным именем
+        List<Task> tasksFromManager = manager.getTasks();
+
+        assertNotNull(tasksFromManager, "Задачи не возвращаются");
+        assertEquals(1, tasksFromManager.size(), "Некорректное количество задач");
+        assertEquals("Test 2", tasksFromManager.get(0).getName(), "Некорректное имя задачи");
+    }
+}
